005_diff_api_perf/pyproject.toml
[tool.poetry]
name = "performance tests"
version = "0.1.0"
description = ""
authors = ["Jozef Sukovsky"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
faker = "^23.2.1"
psycopg2-binary = "^2.9.9"
fastapi = "^0.109.2"
uvicorn = {extras = ["uvloop"], version = "^0.27.1"}
sqlalchemy = "^2.0.27"
gunicorn = "^21.2.0"
django-db-connection-pool = {extras = ["postgresql"], version = "^1.2.4"}
uvloop = "^0.19.0"
bokeh = "^3.4.0"
selenium = "^4.18.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
